requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
gunicorn
psycopg[binary]
ipykernel
dotenv
requests
pandas
numpy
psycopg2
tqdm
pytest
myservice
